admin delete confirmation window  (#167)
* admin delete confirmation window works

* deletes Account nav bar item from menu

* fixing typo in landing page

* admin delete confirmation window works

* deletes Account nav bar item from menu

* fixing typo in landing page

+++ src/components/Organizers/Organizers.jsx
@@ -41,7 +41,9 @@ class Organizers extends Component {
 
     handleDelete(e) {
         const { dispatch, accessToken } = this.props;
-        dispatch(deleteAdmin(e.target.getAttribute('name'), accessToken));
+        if (window.confirm('Please confirm you wish to delete this organizer.')) {
+            dispatch(deleteAdmin(e.target.getAttribute('name'), accessToken));
+        }
     }
 
     handleUpdate(id, index, obj) {
